+++ Kaleido/View/Data/OrderEntityType.swift
@@ -11,7 +11,7 @@ struct OrderEntityType {
     var doer: String
     var note: String
     var pay_method: Int16
-    var product_id: Int32
+    var product_id: Int64
     var remain_money: Int16
     var store_money: Int16
     var total_price: Int16
